Organize common forms and update references
