+++ Client/Forms/Admin/ConfigurationForm.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using PoverkaWinForms.Services;
 using PoverkaWinForms.Forms;
+using PoverkaWinForms.Forms.Common;
 
 namespace PoverkaWinForms.Forms.Admin;
